pyproject.toml
[project]
name = "ai-interview-screener"
version = "0.1.0"
description = "AI Interview Screener Backend (FastAPI)"
authors = [
    {name = "Chaitanya Yadav", email = "jeetkrishna3333@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "python-multipart>=0.0.9",
    "pydantic>=2.7.4",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    "pdfplumber>=0.11.2",
    "python-docx>=1.1.2",
    "psycopg2-binary",
    "google-genai>=1.41.0",
    "sqlalchemy-utils>=0.42.0",
    "twilio>=9.8.3",
]

[tool.uv]

